rename `blob` -> `buffer` utilities file
