+++ packages/shared/src/buffer.ts
